backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
sqlalchemy
bcrypt
python-jose
pydantic
requests
